+++ src/main/java/ru/wert/normic/materials/matlPatches/RoundMatPatchController.java
@@ -46,7 +46,7 @@ public class RoundMatPatchController extends AbstractMatPatchController {
     public void countWeightAndArea() {
         try {
             t = detailController.getCmbxMaterial().getValue().getParamS(); //Диаметр
-            ro = detailController.getCmbxMaterial().getValue().getParamX(); //Масса погонного метра
+            ro = (double) detailController.getCmbxMaterial().getValue().getParamX(); //Масса погонного метра
             paramA = Integer.parseInt(tfA.getText().trim()); //Длина отрезанной заготовки
             paramB = 0;
             wasteRatio = Double.parseDouble(tfWasteRatio.getText().trim());
